Add files via upload
This project attempts to introduce the reader with some general travel patterns of the London Underground Stations 
by considering passenger exits for all stations during service hours. The analysis uses the 'hourlyExits.csv' file
generated by Tfl and identifies two 'peak times' of regular weekdays (in the morning, consisting of Hour07, Hour08, Hour09 and in the evening, consisting of Hour16, Hour17, Hour18). This bi-modal exit distribution is first visualised and further analysed, then used as means to classify stations, according to whether they are 'work-related', 'residential' or 'other' areas. The coursework provides some rudimental justification for the cut-off thresholds (mainly based on plotting the distributions and descriptive statistics). The developed ideas and expectations about stations are then tested on a set of Northern Line Stations.